+++ main.go
@@ -7,7 +7,9 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"strings"
 
+	"github.com/kljensen/snowball/english"
 	"github.com/polisgo2020/search-tarival/index"
 	"github.com/polisgo2020/search-tarival/search"
 )
@@ -42,7 +44,13 @@ func main() {
 		if len(os.Args) < 3 {
 			log.Fatal(errors.New("Search phrase not found"))
 		}
-		searchResult := search.Searching(indexName, pathToStopWords, os.Args[2:])
+		keywords := convertPharaseToKeywords(os.Args[2:], pathToStopWords)
+		index, err := index.ReadIndex(indexName)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		searchResult := search.Searching(index, keywords)
 		output := ""
 		for i, result := range searchResult {
 			output = fmt.Sprintf("%v%v) %v\n", output, i+1, result)
@@ -57,4 +65,19 @@ func main() {
 	default:
 		fmt.Printf("Unknown command.\nRun 'search-tarival help' for usage.")
 	}
+}
+
+func convertPharaseToKeywords(searchPhrase []string, pathToStopWords string) []string {
+	var keywords []string
+
+	mapStopWords := index.CreateStopWordsMap(pathToStopWords)
+
+	for _, keyword := range searchPhrase {
+		keyword = english.Stem(keyword, false)
+		if _, ok := mapStopWords[keyword]; ok {
+			continue
+		}
+		keywords = append(searchPhrase, strings.ToLower(keyword))
+	}
+	return keywords
 }
